Scaffold restaurants
1. Created restaurants scaffold
2. Tested restaurants with curl scripts
3. Changed a data type of phone number in users

+++ db/migrate/20200129194909_add_details_to_users.rb
@@ -2,7 +2,7 @@ class AddDetailsToUsers < ActiveRecord::Migration[5.2]
   def change
     add_column :users, :given_name, :string
     add_column :users, :family_name, :string
-    add_column :users, :phone_number, :integer
+    add_column :users, :phone_number, :bigint
     add_column :users, :food_preferences, :string
   end
 end
